refactor(components): extract generic Calendar primitive from ActivityCalendar
log 固有の upcoming 意味論を tone: default | accent に置き換え、
トリガー + Popover の組み立ては log/_components/log-calendar に移動。
