+++ src/components/calendar/styles.css.ts
@@ -1,29 +1,5 @@
 import { css } from '@styled/css';
 
-// FeedLink と同じ mono テキスト系トーン。metaRow で横に並ぶ前提。
-export const trigger = css({
-  display: 'inline-flex',
-  alignItems: 'center',
-  color: 'fg.muted',
-  cursor: 'pointer',
-  '&[data-hovered]': {
-    color: 'accent.text',
-  },
-});
-
-// Floating surface — sys-bar / quote-share の popover トークンに合わせる。
-export const popover = css({
-  bg: 'bg.canvas',
-  borderWidth: 'default',
-  borderStyle: 'solid',
-  borderColor: 'fg.default',
-});
-
-export const dialog = css({
-  padding: '3',
-  outline: 'none',
-});
-
 export const headerRoot = css({
   display: 'flex',
   alignItems: 'center',
@@ -62,7 +38,7 @@ export const headerCell = css({
 });
 
 // セル下端のドットは ::after で描く（装飾グリフは CSS で描く — JSX に直書きしない）。
-// timeline の記号系を踏襲: 塗り(accent) = upcoming、muted ドット = 過去エントリ。
+// timeline の記号系を踏襲: 塗り(accent) = accent トーン、muted ドット = default トーン。
 export const cell = css({
   position: 'relative',
   display: 'flex',
@@ -89,12 +65,12 @@ export const cell = css({
     borderRadius: 'full',
     bg: 'transparent',
   },
-  '&[data-mark="past"]': {
+  '&[data-tone="default"]': {
     _after: {
       bg: 'fg.muted',
     },
   },
-  '&[data-mark="upcoming"]': {
+  '&[data-tone="accent"]': {
     _after: {
       bg: 'accent.solid',
     },
